# components/gov-chat-backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Government Services API",
      "version": "1.0.0",
      "description": "API documentation for Government Services microservices",
      "contact": {
        "name": "API Support",
        "email": "support@example.com"
      }
    },
    "servers": [
      {
        "url": "/api",
        "description": "API Server"
      }
    ],
...
    "paths": {
...
      "/admin/database/stats": {
        "get": {
          "summary": "Get database statistics",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Database statistics retrieved successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/logs": {
        "get": {
          "summary": "Get system logs",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "query",
              "name": "limit",
              "schema": {
                "type": "integer"
              },
              "description": "Maximum number of logs to return"
            },
            {
              "in": "query",
              "name": "level",
              "schema": {
                "type": "string"
              },
              "description": "Filter logs by level (INFO, WARNING, ERROR)"
            },
            {
              "in": "query",
              "name": "service",
              "schema": {
                "type": "string"
              },
              "description": "Filter logs by service name"
            }
          ],
          "responses": {
            "200": {
              "description": "Logs retrieved successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/logs/rollover": {
        "post": {
          "summary": "Trigger log rollover",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Logs rolled over successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/user-stats": {
        "get": {
          "summary": "Get user statistics",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "User statistics retrieved successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
...
      "/admin/security-scan": {
        "post": {
          "summary": "Run security scan",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Security scan completed successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/security/last-scan": {
        "get": {
          "summary": "Retrieve the last security scan details",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Last security scan details retrieved successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/diagnostics": {
        "post": {
          "summary": "Run system diagnostics",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Diagnostics completed successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/logs/summary": {
        "get": {
          "summary": "Get logs summary by type and service",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "query",
              "name": "date",
              "schema": {
                "type": "string"
              },
              "description": "Date for which to get logs (YYYY-MM-DD)"
            },
            {
              "in": "query",
              "name": "level",
              "schema": {
                "type": "string"
              },
              "description": "Filter by log level"
            }
          ],
          "responses": {
            "200": {
              "description": "Logs summary retrieved successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/logs/search": {
        "get": {
          "summary": "Search logs with filtering",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "query",
              "name": "term",
              "schema": {
                "type": "string"
              },
              "description": "Search term"
            },
            {
              "in": "query",
              "name": "level",
              "schema": {
                "type": "string"
              },
              "description": "Filter by log level"
            },
            {
              "in": "query",
              "name": "service",
              "schema": {
                "type": "string"
              },
              "description": "Filter by service name"
            },
            {
              "in": "query",
              "name": "dateRange",
              "schema": {
                "type": "string",
                "enum": [
                  "today",
                  "yesterday",
                  "week",
                  "month",
                  "custom"
                ]
              },
              "description": "Date range preset"
            },
            {
              "in": "query",
              "name": "startDate",
              "schema": {
                "type": "string"
              },
              "description": "Custom start date (YYYY-MM-DD)"
            },
            {
              "in": "query",
              "name": "endDate",
              "schema": {
                "type": "string"
              },
              "description": "Custom end date (YYYY-MM-DD)"
            }
          ],
          "responses": {
            "200": {
              "description": "Search completed successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/logs/debug-yesterday": {
        "get": {
          "summary": "Debug logs for yesterday to diagnose issues",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Debug information retrieved successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/database-operations/reindex": {
        "post": {
          "summary": "Reindex database",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Database reindexed successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/database-operations/backup": {
        "post": {
          "summary": "Backup database",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Database backed up successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/database-operations/optimize": {
        "post": {
          "summary": "Optimize database",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Database optimized successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/users/search": {
        "get": {
          "summary": "Search users with filtering",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "query",
              "name": "term",
              "schema": {
                "type": "string"
              },
              "description": "Search term"
            },
            {
              "in": "query",
              "name": "field",
              "schema": {
                "type": "string",
                "enum": [
                  "all",
                  "name",
                  "email",
                  "role"
                ]
              },
              "description": "Field to search (all, name, email, role)"
            },
            {
              "in": "query",
              "name": "limit",
              "schema": {
                "type": "integer"
              },
              "description": "Maximum number of users to return"
            },
            {
              "in": "query",
              "name": "offset",
              "schema": {
                "type": "integer"
              },
              "description": "Offset for pagination"
            }
          ],
          "responses": {
            "200": {
              "description": "Search completed successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      }
    }
  }